feat: Add createBranch command for creating new branches
This commit adds a new command called createBranch to the extension. The createBranch command allows users to create a new branch in the Git repository. It prompts the user for the name of the new branch and the source branch from which to create the new branch. If both branch names are provided, the command uses the simple-git library to checkout the source branch and then create a new local branch with the specified name. If successful, it displays an information message indicating that the branch has been created. If there is an error during the process, it displays an error message.

This commit also includes changes to the package.json file, updating the version number to 0.0.15 and adding the createBranch command to the extension's commands and menus.

Issue #123